third

+++ src/main/java/com/service/AddressService.java
@@ -1,41 +1,42 @@
 package com.service;
 
-import com.dao.AddressDAO;
-import com.model.Address;
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import com.dao.AddressDAO;
+import com.model.Address;
 
 @Service
 public class AddressService {
 
-	   @Autowired
-    private final AddressDAO addressDAO;
+    @Autowired
+    private AddressDAO addressDAO;
 
- 
-    public AddressService(AddressDAO addressDAO) {
-        this.addressDAO = addressDAO;
-    }
-
-    // Save a new address
+    // Create or update an address
     public Address saveAddress(Address address) {
         return addressDAO.save(address);
     }
 
-    // Get all addresses
+    // Retrieve an address by ID
+    public Address getAddressById(Integer addressId) {
+        Optional<Address> address = addressDAO.findById(addressId);
+        return address.orElse(null); // Return the address if found, otherwise null
+    }
+
+    // Retrieve all addresses
     public List<Address> getAllAddresses() {
         return addressDAO.findAll();
     }
 
-    // Find address by ID
-    public Optional<Address> getAddressById(Integer addressId) {
-        return addressDAO.findById(addressId);
-    }
-
-    // Delete address by ID
-    public void deleteAddress(Integer addressId) {
-        addressDAO.deleteById(addressId);
+    // Delete an address by ID
+    public boolean deleteAddress(Integer addressId) {
+        if (addressDAO.existsById(addressId)) {
+            addressDAO.deleteById(addressId);
+            return true;
+        }
+        return false; // Address not found
     }
 }
